Move source code into src folder
