--- lib/phuNames.js
@@ -1,38 +1,0 @@
-const phuNames = [
-	'HALIBURTON, KAWARTHA, PINE RIDGE',
-	'HALTON REGION',
-	'HASTINGS & PRINCE EDWARD COUNTIES',
-	'HURON PERTH',
-	'KINGSTON, FRONTENAC, LENNOX & ADDINGTON',
-	'LAMBTON COUNTY',
-	'LEEDS, GRENVILLE AND LANARK DISTRICT',
-	'MIDDLESEX-LONDON',
-	'NIAGARA REGION',
-	'NORTH BAY PARRY SOUND DISTRICT',
-	'NORTHWESTERN',
-	'OXFORD ELGIN-ST.THOMAS',
-	'PEEL REGION',
-	'PETERBOROUGH COUNTY-CITY',
-	'PORCUPINE',
-	'RENFREW COUNTY AND DISTRICT',
-	'SIMCOE MUSKOKA DISTRICT',
-	'SUDBURY AND DISTRICT',
-	'THUNDER BAY DISTRICT',
-	'TIMISKAMING',
-	'TORONTO',
-	'WATERLOO REGION',
-	'WELLINGTON-DUFFERIN-GUELPH',
-	'WINDSOR-ESSEX COUNTY',
-	'YORK REGION',
-	'ALGOMA DISTRICT',
-	'BRANT COUNTY',
-	'CHATHAM-KENT',
-	'CITY OF HAMILTON',
-	'CITY OF OTTAWA',
-	'DURHAM REGION',
-	'EASTERN ONTARIO',
-	'GREY BRUCE',
-	'HALDIMAND-NORFOLK',
-];
-
-export default phuNames;
