+++ src/chartConfig/ontario/hospitalized.js
