+++ src/chartConfig/ontario/totalDeaths.js
