+++ src/chartConfig/vaccinations/totalVaccinated.js
