+++ src/components/SmallContentContainer.js
